\name{print-methods}
\alias{print.ddct}
\alias{print.ct}

\title{print methods}
\description{print methods}
\usage{
\method{print}{ddct}(x, digits = 3, ...)
\method{print}{ct}(x, digits = 3, ...)
}
\arguments{
  \item{x}{An object of class qpcrmacro}
  \item{digits}{number of digits}
  \item{\dots}{\dots}
}

\author{Daniel Gerhard}

\seealso{\code{\link{qpcr_nlme}}, \code{\link{qpcr_nlme_formula}}}

\keyword{methods}
